+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Changelog
 * Add new method :meth:`common.conda_env`.
 * Add new argument ``meta_only`` to :meth:`pyvcf.VcfFrame.from_file` method.
 * Add new argument ``nrows`` to :meth:`pyvcf.VcfFrame.from_file` method.
+* Add new method :meth:`pyvcf.VcfFrame.extract_info`.
 
 0.20.0 (2021-08-07)
 -------------------